add utils subproject
