+++ targetmine-bio-sources/utils/src/main/resources/umls_keys.properties
@@ -1,0 +1,9 @@
+DataSet.key_name = name
+DataSource.key_name = name
+DiseaseTerm.key_name = identifier
+MeshTerm.key_name = identifier
+UMLSDisease.key_name = identifier
+SOTerm.key_name_ontology = name, ontology
+Organism.key_taxonid = taxonId
+Ontology.key_name = name
+Publication.key_pubmedid = pubMedId
